started added bottom nav bar
