+++ app/src/main/res/drawable/ic_home.xml
@@ -1,0 +1,13 @@
+<vector xmlns:android="http://schemas.android.com/apk/res/android"
+    android:width="24dp"
+    android:height="24dp"
+    android:viewportWidth="24"
+    android:viewportHeight="24">
+  <group>
+    <clip-path
+        android:pathData="M0,0h24v24h-24z"/>
+    <path
+        android:pathData="M12,3L4,9V21H9V14H15V21H20V9L12,3Z"
+        android:fillColor="#FFFFFF"/>
+  </group>
+</vector>
