+++ app/src/main/res/menu/bottom_nav_menu.xml
@@ -1,0 +1,10 @@
+<?xml version="1.0" encoding="utf-8"?>
+<menu xmlns:android="http://schemas.android.com/apk/res/android">
+    <item android:id="@+id/bottom_menu_home"
+        android:title="@string/home_view"
+        android:icon="@drawable/ic_home"/>
+
+    <item android:id="@+id/bottom_menu_history"
+        android:title="@string/history_view"
+        android:icon="@drawable/ic_history"/>
+</menu>
