+++ app/src/main/res/values/strings.xml
@@ -2,6 +2,10 @@
     <string name="app_name" translatable="false">HeartRateMonitor</string>
     <string name="about_view">About</string>
     <string name="settings_view">Settings</string>
+    <string name="home_view" translatable="false">Home</string>
+    <string name="history_view">History</string>
+
+    <string name="heart_rate_view_title">Measure heart rate</string>
 
     <string name="text_heart_rate_idle">Press button to start measuring</string>
     <string name="text_heart_rate" translatable="false">%1$s bpm</string>
